Fix project setup (#6)
The project was created using the yeoman generator, but this way the
support for multiple projects or Visual Studio is aweful. I fixed the
project structure so it works better with Visual Studio.

I also upped the versions to be compatible with RC2 of dotnet and aspnet
core. This will cause some changes (like using the dotnet tool instead of
dnx).